Added files for new DS PyLyrtechDLLRF.
This is the device version 1.0 of the PyLyrtechDLLRF device which runs
in windows (written in python).

Tis is just a python refactoring of the current LyrtechDLLRF device
written in C, that controls the Nutaq Lyrtech VHS-ADC/DAC FPGA in charge
of the LLRF algorithm.

It contains basic functions to initialize, configurwe defaults, check
the status and read/write settings in the custom registers.